unsigned ints

+++ hash.h
@@ -29,6 +29,7 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #define BOOST_DATE_TIME_NO_LIB
 #include <h3.h>
+#include <vector>
 #include <boost/interprocess/allocators/allocator.hpp>
 #include <boost/interprocess/managed_shared_memory.hpp>
 
@@ -45,7 +46,7 @@ typedef boost::interprocess::allocator<void, segment_manager_t> void_allocator;
 class hash_function {
    public:
     // Max object size in bytes that we will receive
-    const static size_t max_obj_size = 36;
+    const static int max_obj_size = 36;
 
     hash_function(size_t seed);
